commit on 'master'

+++ index.js
@@ -18,22 +18,23 @@ bot.on("message",message=>{
             world.send(message.author.username+" made the suggestion\n"+suggestion).then(()=>message.channel.send("Thank you for your suggestion"));
         }
     }
-id(message.channel.id == "725041470289215629"){
-    if(message.content.startsWith("-poll")){
-        var words = message.content.split(" ");
-        if(Number(words[1])){
-            for(var i = 1; i <= Number(words[1]); i++){
-                message.react(array[i]);
-            }
+    if(message.channel.id == "725041470289215629"){
+        if(message.content.startsWith("-poll")){
+            var words = message.content.split(" ");
+            if(Number(words[1])){
+                for(var i = 1; i <= Number(words[1]); i++){
+                    message.react(array[i]);
+                }
+            }else{
+                message.react("👍").then(()=>{
+                    message.react("👎")
+                });
+             }
         }else{
-            message.react("👍").then(()=>{
-                message.react("👎")
-            });
-        }else{
-message.delete();
-}
+            message.delete();
+        }
     }
-}
+        
     if(message.content.startsWith("-lua")){
         var message_to_send = message.content.slice(4);
         message.delete();
